+++ api/src/main/java/org/toy_project/account/application/port/in/GetAccountUseCase.java
@@ -1,0 +1,10 @@
+package org.toy_project.account.application.port.in;
+
+import org.springframework.stereotype.Component;
+import org.toy_project.account.adapter.out.persistence.Account;
+
+@Component
+public interface GetAccountUseCase {
+
+    Account getAccount(Long id);
+}
