+++ api/src/main/java/org/toy_project/account/domain/AccountImageEntity.java
@@ -1,0 +1,31 @@
+package org.toy_project.account.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "account_image")
+public class AccountImageEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "idx")
+    private Long id;
+
+    @Column(name = "type")
+    private String type;
+
+    @Column(name = "image_url")
+    private String imageUrl;
+
+    @Column(name = "visible")
+    private boolean visible;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "account_id")
+    private AccountEntity accountEntity;
+}
