+++ api/src/main/java/org/toy_project/post/application/service/PostService.java
@@ -1,0 +1,30 @@
+package org.toy_project.post.application.service;
+
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.toy_project.post.application.port.in.GetPostUseCase;
+import org.toy_project.post.adapter.out.persistence.Post;
+import org.toy_project.post.application.port.out.LoadPostPort;
+
+@Service
+@RequiredArgsConstructor
+class PostService implements GetPostUseCase {
+
+    private final LoadPostPort loadPostPort;
+
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<Post> getPostList() {
+        return loadPostPort.loadPostList();
+    }
+
+
+    @Override
+    @Transactional(readOnly = true)
+    public Post getPost(Long id) {
+        return loadPostPort.loadPost(id);
+    }
+}
